+++ src/Vortice.XAudio2/Fx/ReverbParameters.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Amer Koleci and contributors.
+// Distributed under the MIT license. See the LICENSE file in the project root for more information.
+
+using SharpGen.Runtime;
+
+namespace Vortice.XAudio2.Fx
+{
+    public partial struct ReverbParameters
+    {
+        
+    }
+}